requirements.txt
pyyaml
matplotlib
aiohttp[speedups]
beautifulsoup4
lxml
yarl

--extra-index-url https://google-coral.github.io/py-repo/
tflite_runtime
